+++ src/service/TRex/src/netstandard/VSS.TRex.Filters.Models/FilteredValueAssignmentContext.cs
@@ -5,7 +5,7 @@ namespace VSS.TRex.Filters.Models
 {
     /// <summary>
     /// FilteredValueAssignmentContext provides a context for collecting filtered values from production data in the 
-    /// process of applying spatial, temporal and attributal filtering constraints.
+    /// process of applying spatial, temporal and attribute filtering constraints.
     /// </summary>
     public class FilteredValueAssignmentContext
     {
@@ -14,15 +14,15 @@ namespace VSS.TRex.Filters.Models
         /// </summary>
         public struct ProbePoint
         {
-            public double XOffset, YOffset;
+            public float XOffset, YOffset;
 
-            public ProbePoint(double xOffset, double yOffset)
+            public ProbePoint(float xOffset, float yOffset)
             {
                 XOffset = xOffset;
                 YOffset = yOffset;
             }
 
-            public void SetOffsets(double xOffset, double yOffset)
+            public void SetOffsets(float xOffset, float yOffset)
             {
                 XOffset = xOffset;
                 YOffset = yOffset;
@@ -39,9 +39,9 @@ namespace VSS.TRex.Filters.Models
 
         /// <summary>
         /// ProbePositions is used to store the real world positions used to probe into
-        /// cells when computing floating type seived bitmasks during subgrid querying.
+        /// cells when computing floating type sieved bit masks during subgrid querying.
         /// The positions are stored as offsets from the real world origin of the subgrid
-        /// As many queries do not need access to probe positions the arfray is created
+        /// As many queries do not need access to probe positions the array is created
         /// only if InitialiseProbePositions() is called
         /// </summary>
         public ProbePoint[,] ProbePositions;
